+++ yash-env/src/system/real.rs
@@ -17,6 +17,7 @@
 //! Implementation of `System` that actually interacts with the system.
 
 mod errno;
+mod file_system;
 mod open_flag;
 mod signal;
 
@@ -30,13 +31,13 @@ use super::Env;
 use super::Errno;
 use super::FdFlag;
 use super::FdSet;
-use super::FileStat;
 use super::Gid;
 use super::Mode;
 use super::OfdAccess;
 use super::OpenFlag;
 use super::Result;
 use super::SigmaskOp;
+use super::Stat;
 use super::System;
 use super::TimeSpec;
 use super::Times;
@@ -184,17 +185,23 @@ impl RealSystem {
 }
 
 impl System for RealSystem {
-    fn fstat(&self, fd: Fd) -> Result<FileStat> {
-        Ok(nix::sys::stat::fstat(fd.0)?)
+    fn fstat(&self, fd: Fd) -> Result<Stat> {
+        let mut stat = MaybeUninit::<nix::libc::stat>::uninit();
+        unsafe { nix::libc::fstat(fd.0, stat.as_mut_ptr()) }.errno_if_m1()?;
+        Ok(Stat::from_raw(&stat))
     }
 
-    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<FileStat> {
+    fn fstatat(&self, dir_fd: Fd, path: &CStr, follow_symlinks: bool) -> Result<Stat> {
         let flags = if follow_symlinks {
-            AtFlags::empty()
+            0
         } else {
-            AtFlags::AT_SYMLINK_NOFOLLOW
+            nix::libc::AT_SYMLINK_NOFOLLOW
         };
-        Ok(nix::sys::stat::fstatat(dir_fd.0, path, flags)?)
+
+        let mut stat = MaybeUninit::<nix::libc::stat>::uninit();
+        unsafe { nix::libc::fstatat(dir_fd.0, path.as_ptr(), stat.as_mut_ptr(), flags) }
+            .errno_if_m1()?;
+        Ok(Stat::from_raw(&stat))
     }
 
     fn is_executable_file(&self, path: &CStr) -> bool {