+++ yash-semantics/src/redir.rs
@@ -96,10 +96,10 @@ use yash_env::semantics::ExitStatus;
 use yash_env::semantics::Field;
 use yash_env::system::Errno;
 use yash_env::system::FdFlag;
+use yash_env::system::FileType;
 use yash_env::system::Mode;
 use yash_env::system::OfdAccess;
 use yash_env::system::OpenFlag;
-use yash_env::system::SFlag;
 use yash_env::Env;
 use yash_env::System;
 use yash_quote::quoted;
@@ -315,8 +315,9 @@ fn open_file_noclobber(env: &mut Env, path: Field) -> Result<(FdSpec, Location),
     // Okay, it seems there is an existing file. Try opening it.
     match system.open(&path, OfdAccess::WriteOnly, EnumSet::empty(), MODE) {
         Ok(fd) => {
-            let is_regular = matches!(system.fstat(fd), Ok(stat)
-                    if SFlag::from_bits_truncate(stat.st_mode) & SFlag::S_IFMT == SFlag::S_IFREG);
+            let is_regular = system
+                .fstat(fd)
+                .is_ok_and(|stat| stat.r#type == FileType::Regular);
             if is_regular {
                 // We opened the FD without the O_CREAT flag, so somebody else
                 // must have created this file. Failure.